feat(container): add tests to pnr
ref: MANAGER-14778
