+++ packages/manager/apps/container/src/container/nav-reshuffle/sidebar/ProjectSelector/PciProject.ts
@@ -1,0 +1,13 @@
+export type PciProject = {
+  access: string;
+  creationDate: Date;
+  description: string;
+  expiration: Date;
+  manualQuota: boolean;
+  orderId: unknown;
+  planCode: string;
+  projectName: string;
+  project_id: string;
+  status: string;
+  unleash: boolean;
+};
