+++ packages/manager/apps/container/src/container/nav-reshuffle/sidebar/ProjectSelector/ProjectSelector.tsx
@@ -1,19 +1,6 @@
 import React, { useEffect, useState } from 'react';
 import Select from 'react-select';
-
-export type PciProject = {
-  access: string;
-  creationDate: Date;
-  description: string;
-  expiration: Date;
-  manualQuota: boolean;
-  orderId: unknown;
-  planCode: string;
-  projectName: string;
-  project_id: string;
-  status: string;
-  unleash: boolean;
-};
+import { PciProject } from './PciProject';
 
 const getProjectOption = (option: Record<string, any>): JSX.Element => {
   return (
@@ -156,6 +143,7 @@ const ProjectSelector: React.FC<ComponentProps<Props>> = ({
         menuPosition={'absolute'}
         onMenuOpen={() => onMenuOpen && onMenuOpen()}
         value={value}
+        data-testid="project-selector"
         onChange={(option) => {
           if (option.new) {
             onProjectCreate();
